Programming Assignment 3 Part B
Given a string s which represents an expression, evaluate this expression and return its value. 
The integer division should truncate toward zero. You may assume that the given expression is always valid. 
Note: You are NOT allowed to use any built-in function, or external library, which evaluates strings as mathematical expressions, such as eval().
Constraints:
s consists of integers and operators  '+', '-', '*', and  '/'  separated by some number of spaces. No other characters and assume the expression is valid
operator precedence 
-, +, *, /   (low to high) (note: this is slightly different from math; - and + are NOT in the same level; so do * and /)
s represents a valid expression. All the integers in the expression are non-negative integers in the range [0, 2^31 - 1].